[Ubuntu] Decompress the image
This was insanely complex for how simple it was. TLDR is that bazel
downloads a file that does not have `.xz` in the extention. This means
xz will refuse to run on it. We try to make a symlink to the file so
that it's named with a `.xz` but bazel doesn't like this. We instead
find the original base file using `readlink` in order to create the
symlink. This appeases the bazel gods and we can return to a life of
peace and prosperity knowing our image is decompressed.

Signed-off-by: Greg Balke <gbalke@berkeley.edu>